rename: hooks/feature를 hooks/queries로 면경
